requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.35.0
gspread
oauth2client
